DDF-4180 Moving bundle
DDF-4180 Exporting api package

DDF-4180 Addressing codacy warnings

DDF-4180 Update constants

DDF-4035 Constraining java.ext.dirs to java home in external Solr (#3814) (#3819)


DDF-4184 Resolve Sonar Warnings (#3811)


DDF-4193 make the installer set system.hostname to the given hostname when dev=true. (#3832)


DDF-4198 Fixed a typo in the property for 'Internal Default Port' (#3838)


DDF-4180 Updating coverage

DDF-4173 Adding MutlLine/MultiPoly Rendering Support for OL (#3795)

* DDF-4173 Adding MutlLine/MultiPoly Rendering Support for OL

* DDF-4173 Cleanup

* DDF-4173 Fixing invalid mapping syntax

DDF-4189 Fix banner z-index. (#3822)


DDF-4179 Map pan & zoom by geographic place name appears to search forever and does not reorient the camera (#3805)


DDF-4183 Single quoted property name support in cql.js (#3813)

* DDF-4183 Single quoted property name support in cql.js

* DDF-4183 Added cql.js unit test

DDF-4161 Fix registry date field validation error (#3771)

- No longer require time field when validating.

DDF-4176 Update ConfluenceFilterDelegate to handle propertyIsGreaterThan (#3802)

- Also fix and issue with the TagsFilterDelegate 'not' implementation

DDF-4190 Added docs on selecting transformer for CSW Insert (#3825) (#3846)


DDF-3938 Update to Solr 7 style config and schema

DDF-3938 Use Solr Real Time Get for ID queries

DDF-3938 Remove dependency on text_ws

DDF-3938 Increase auto commit max time

DDF-3938 Protect getById from 414 URI Too Long

DDF-3938 Remove code related to *_txt_ws*

DDF-3938 Use terms query parser for ID queries

DDF-3938 Use docValues to improve sorting performance

DDF-3938 Match max boolean clauses with solrconfig.xml

DDF-3938 Fix validation itest queries and assertions

DDF-3938 Replace SpatialRecursivePrefixTreeFieldType with RptWithGeometrySpatialField

DDF-3938 Replace Trie fields with Point fields

DDF-3938 Remove unused parts of schema.xml

DDF-3938 Replace deprecated SynonymFilterFactory

DDF-3938 Use Solr 7 DocValues to replace sort key fields

DDF-3938 Set auto hard commit maximum size to protect tlog size

DDF-4205 Typo (#3845)


DDF-4192 Add validation to search forms (#3833)

* DDF-4192 Add validation to search forms

* DDF-4192 fixes

* DDF-4192 addressing comments

* DDF-4192 formatting

DDF-4185 Allow custom stopping of execution for uploading lists (#3815)

* DDF-4185 Added new exception

* DDF-4185 Fixing message errors and removing comments

* DDF-4185 Addressed PR comments Round 1

* DDF-4185 Addressed PR comments round 2

* GSR-4185 Added better JavaDocs

* GSR-4185 Ran formatter

* DDF-4185 Refactored Splitter Interfaces and Exception class into the API bundle

* DDF-4185 Moved ServiceProperties into catalog-ui-search-api bundle

* DDF-4185 Ran formatter

* DDF-4185 Fixed typo

* DDF-4185 Adjusted pom file for catalog-ui-search-api

Merge
